+++ app_backend/app/routers/user.py
@@ -6,20 +6,6 @@ from database import db
 router = APIRouter(prefix="/user")
 
 
-"""
-    UserLogin model keeps track of the user login information
-    username: str
-        The username of the user
-    email: str
-        The email of the user
-    password: str
-        The password of the user
-"""
-class User(BaseModel):
-    name: str
-    email: str
-    password: str
-
 
 """
     getuser implements the route /login/
